refactor: Basket agreement useMutations
Moved the react query useMutation hooks to basket view and agreement modal levels
