+++ src/components/AgreementModal/AgreementModal.js
@@ -1,12 +1,17 @@
 import PropTypes from 'prop-types';
+import { useMutation } from 'react-query';
+import { useLocation, useHistory } from 'react-router-dom';
 import { Field } from 'react-final-form';
 import arrayMutators from 'final-form-arrays';
 import { FormattedMessage } from 'react-intl';
+
+import { useOkapiKy } from '@folio/stripes/core';
 
 import {
   composeValidators,
   requiredValidator,
   useAsyncValidation,
+  getRefdataValuesByDesc,
 } from '@folio/stripes-erm-components';
 
 import {
@@ -18,7 +23,11 @@ import {
 } from '@folio/stripes/components';
 
 import { FormModal } from '@k-int/stripes-kint-components';
+
 import { validationEndPoint } from '../../constants';
+import { AGREEMENTS_ENDPOINT } from '../../constants/endpoints';
+import { useBasket, useAgreementsRefdata } from '../../hooks';
+import { urls } from '../utilities';
 
 const propTypes = {
   showModal: PropTypes.bool,
@@ -27,22 +36,49 @@ const propTypes = {
   agreementStatusValues: PropTypes.arrayOf(PropTypes.object),
   handleAgreement: PropTypes.func,
 };
+const [AGREEMENT_STATUS] = ['SubscriptionAgreement.AgreementStatus'];
 
-const AgreementModal = ({
-  showModal,
-  hideModal,
-  selectedItems,
-  agreementStatusValues,
-  handleAgreement,
-}) => {
+const AgreementModal = ({ showModal, hideModal, selectedItems }) => {
+  const location = useLocation();
+  const history = useHistory();
+  const ky = useOkapiKy();
+  const refdata = useAgreementsRefdata({
+    desc: [AGREEMENT_STATUS],
+  });
+
+  const { basket } = useBasket();
+
   const validateAsyncBackend = useAsyncValidation(
     'ui-agreements',
     validationEndPoint.AGREEMENTPATH
   );
 
-  const submitAgreement = async (values) => {
-    await handleAgreement(selectedItems, values);
-    hideModal();
+  const { mutateAsync: postAgreement } = useMutation(
+    [
+      AGREEMENTS_ENDPOINT,
+      'ui-agreements',
+      'AgreementCreateRoute',
+      'createAgreement',
+    ],
+    (payload) => ky
+      .post(AGREEMENTS_ENDPOINT, { json: payload })
+      .json()
+      .then(({ id }) => {
+        history.push(`${urls.agreementView(id)}${location.search}`);
+      })
+  );
+
+  const handleAddToNewAgreement = async (values) => {
+    const submitValues = {
+      name: values?.name,
+      agreementStatus: values?.agreementStatus,
+      periods: [{ startDate: values?.startDate }],
+      items: selectedItems
+        .split(',')
+        .map((index) => ({ resource: basket[parseInt(index, 10)] }))
+        .filter((line) => line.resource),
+    };
+    await postAgreement(submitValues);
   };
 
   return (
@@ -50,10 +86,12 @@ const AgreementModal = ({
       modalProps={{
         onClose: hideModal,
         open: showModal,
-        label: <FormattedMessage id="ui-agreements.agreements.createAgreement" />,
+        label: (
+          <FormattedMessage id="ui-agreements.agreements.createAgreement" />
+        ),
       }}
       mutators={arrayMutators}
-      onSubmit={submitAgreement}
+      onSubmit={handleAddToNewAgreement}
     >
       <Row>
         <Col xs={4}>
@@ -77,7 +115,10 @@ const AgreementModal = ({
               return (
                 <Select
                   {...input}
-                  dataOptions={agreementStatusValues}
+                  dataOptions={getRefdataValuesByDesc(
+                    refdata,
+                    AGREEMENT_STATUS
+                  )}
                   error={meta && meta.touched && meta.error}
                   id="edit-agreement-status"
                   label={
